Add a default settings implementation
Closes #10
