+++ samples/Raygun.Owin.Samples.NancyFX/Web.config
@@ -3,7 +3,7 @@
     <appSettings>
         <add key="raygun:apiKey" value="YOUR_APP_API_KEY" />
         <add key="raygun:tags" value="nancy, sample" />
-        <add key="raygun:applicationVersion" value="1.2.3.4-beta2"/>
+        <add key="raygun:applicationVersion" value="1.2.3.4-beta2" />
     </appSettings>
 
     <system.web>
